requirements.txt
attrs==22.2.0
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==39.0.0
exceptiongroup==1.1.0
iniconfig==2.0.0
logger==1.4
packaging==23.0
paramiko==2.12.0
Pint==0.18
pluggy==1.0.0
pycparser==2.21
pyfrc==2023.0.0
PyNaCl==1.5.0
pynetconsole==2.0.4
pynetworktables==2021.0.0
pyntcore==2023.1.1.0
pytest==7.2.0
pytest-reraise==2.1.2
robotpy==2023.1.1
robotpy-apriltag==2023.1.1.0
robotpy-commands-v2==2023.1.1.0
robotpy-ctre==2023.0.0
robotpy-hal==2023.1.1.0
robotpy-halsim-gui==2023.1.1.0
robotpy-installer==2023.0.1
robotpy-photonvision==2023.1.1b7
robotpy-rev==2023.1.1
robotpy-toolkit-7407==0.5.21
robotpy-wpilib-utilities==2023.0.0
robotpy-wpimath==2023.1.1.0
robotpy-wpinet==2023.1.1.0
robotpy-wpiutil==2023.1.1.0
six==1.16.0
tomli==2.0.1
wpilib==2023.1.1.0
